refactor: separate constant file
