--- src/js/constants.js
@@ -1,18 +1,0 @@
-export const CAR_NAME_LENGTH = {
-  MIN: 1,
-  MAX: 5,
-};
-export const RANDOM_RANGE = {
-  MIN: 0,
-  MAX: 9,
-};
-export const MIN_ADVANCE_VALUE = 4;
-export const MIN_RACE_COUNT = 1;
-export const MIN_CAR_COUNT = 2;
-
-export const EXCEPTIONS = {
-  INVALID_CAR_NAMES: `조건에 맞게 중복 없이 ${MIN_CAR_COUNT}대 이상을 입력해주세요.`,
-  INVALID_RACE_COUNT: `${MIN_RACE_COUNT} 이상의 정수를 입력해주세요.`,
-};
-
-export const KEYBOARD_ENTER = 13;
