+++ src/js/constants/events.js
@@ -1,0 +1,1 @@
+export const KEYBOARD_ENTER = 13;
