+++ src/js/constants/games.js
@@ -1,0 +1,11 @@
+export const CAR_NAME_LENGTH = {
+  MIN: 1,
+  MAX: 5,
+};
+export const RANDOM_RANGE = {
+  MIN: 0,
+  MAX: 9,
+};
+export const MIN_ADVANCE_VALUE = 4;
+export const MIN_RACE_COUNT = 1;
+export const MIN_CAR_COUNT = 2;
