+++ src/js/controller/racingCount/checkFunctions.js
@@ -1,4 +1,4 @@
-import { MIN_RACE_COUNT } from "../../constants.js";
+import { MIN_RACE_COUNT } from "../../constants/games.js";
 
 const isInteger = input => {
   if (Number.isNaN(input) || Number(input) % 1 !== 0) {
